refactor: remove onClose responsibility from Validate Subscribers Form.tsx
